v1.7.2 #1 Added "Write SWS actions to sws_actions.csv"
Updated whatsnew.txt with historical information.  Devs: Please update whatsnew.txt with all your changes.  See:
http://code.google.com/p/sws-extension/wiki/WhatsNewFormat

+++ Xenakios/MixerActions.cpp
@@ -958,7 +958,7 @@ void DoSetRenderSpeedToRealtime2(COMMAND_T*)
 	int sz=0; int *renderspeedmode = (int *)get_config_var("workrender",&sz);
     if (sz==sizeof(int) && renderspeedmode) 
 	{ 
-		bitset<32> blah((ULONG)*renderspeedmode);	
+		bitset<32> blah(*renderspeedmode);	
 		blah.set(3);
 		*renderspeedmode=blah.to_ulong();
 	}
@@ -971,7 +971,7 @@ void DoSetRenderSpeedToNonLim(COMMAND_T*)
 	int sz=0; int *renderspeedmode = (int *)get_config_var("workrender",&sz);
     if (sz==sizeof(int) && renderspeedmode) 
 	{ 
-		bitset<32> blah((ULONG)*renderspeedmode);	
+		bitset<32> blah(*renderspeedmode);	
 		blah.reset(3);
 		*renderspeedmode=blah.to_ulong();
 	}
